MM-13960 Removed some old combining logic and moved the rest
