+++ src/utils/post_list.js
@@ -155,3 +155,139 @@ export function makeCombineUserActivityPosts() {
         },
     );
 }
+
+export const postTypePriority = {
+    [Posts.POST_TYPES.JOIN_TEAM]: 0,
+    [Posts.POST_TYPES.ADD_TO_TEAM]: 1,
+    [Posts.POST_TYPES.LEAVE_TEAM]: 2,
+    [Posts.POST_TYPES.REMOVE_FROM_TEAM]: 3,
+    [Posts.POST_TYPES.JOIN_CHANNEL]: 4,
+    [Posts.POST_TYPES.ADD_TO_CHANNEL]: 5,
+    [Posts.POST_TYPES.LEAVE_CHANNEL]: 6,
+    [Posts.POST_TYPES.REMOVE_FROM_CHANNEL]: 7,
+    [Posts.POST_TYPES.PURPOSE_CHANGE]: 8,
+    [Posts.POST_TYPES.HEADER_CHANGE]: 9,
+    [Posts.POST_TYPES.JOIN_LEAVE]: 10,
+    [Posts.POST_TYPES.DISPLAYNAME_CHANGE]: 11,
+    [Posts.POST_TYPES.CONVERT_CHANNEL]: 12,
+    [Posts.POST_TYPES.CHANNEL_DELETED]: 13,
+    [Posts.POST_TYPES.ADD_REMOVE]: 14,
+    [Posts.POST_TYPES.EPHEMERAL]: 15,
+};
+
+export function comparePostTypes(a, b) {
+    return postTypePriority[a.postType] - postTypePriority[b.postType];
+}
+
+function extractUserActivityData(userActivities) {
+    const messageData = [];
+    const allUserIds = [];
+    const allUsernames = [];
+
+    Object.entries(userActivities).forEach(([postType, values]) => {
+        if (
+            postType === Posts.POST_TYPES.ADD_TO_TEAM ||
+            postType === Posts.POST_TYPES.ADD_TO_CHANNEL ||
+            postType === Posts.POST_TYPES.REMOVE_FROM_CHANNEL
+        ) {
+            Object.keys(values).map((key) => [key, values[key]]).forEach(([actorId, users]) => {
+                if (Array.isArray(users)) {
+                    throw new Error('Invalid Post activity data');
+                }
+                const {ids, usernames} = users;
+                messageData.push({postType, userIds: [...usernames, ...ids], actorId});
+                if (ids.length > 0) {
+                    allUserIds.push(...ids);
+                }
+
+                if (usernames.length > 0) {
+                    allUsernames.push(...usernames);
+                }
+                allUserIds.push(actorId);
+            });
+        } else {
+            if (!Array.isArray(values)) {
+                throw new Error('Invalid Post activity data');
+            }
+            messageData.push({postType, userIds: values});
+            allUserIds.push(...values);
+        }
+    });
+
+    messageData.sort(comparePostTypes);
+
+    function reduceUsers(acc, curr) {
+        if (!acc.includes(curr)) {
+            acc.push(curr);
+        }
+        return acc;
+    }
+
+    return {
+        allUserIds: allUserIds.reduce(reduceUsers, []),
+        allUsernames: allUsernames.reduce(reduceUsers, []),
+        messageData,
+    };
+}
+
+export function combineUserActivitySystemPost(systemPosts = []) {
+    if (systemPosts.length === 0) {
+        return null;
+    }
+
+    const userActivities = systemPosts.reduce((acc, post) => {
+        const postType = post.type;
+        let userActivityProps = acc;
+        const combinedPostType = userActivityProps[postType];
+
+        if (
+            postType === Posts.POST_TYPES.ADD_TO_TEAM ||
+            postType === Posts.POST_TYPES.ADD_TO_CHANNEL ||
+            postType === Posts.POST_TYPES.REMOVE_FROM_CHANNEL
+        ) {
+            const userId = post.props.addedUserId || post.props.removedUserId;
+            const username = post.props.addedUsername || post.props.removedUsername;
+            if (combinedPostType) {
+                if (Array.isArray(combinedPostType[post.user_id])) {
+                    throw new Error('Invalid Post activity data');
+                }
+                const users = combinedPostType[post.user_id] || {ids: [], usernames: []};
+                if (userId) {
+                    if (!users.ids.includes(userId)) {
+                        users.ids.push(userId);
+                    }
+                } else if (username && !users.usernames.includes(username)) {
+                    users.usernames.push(username);
+                }
+                combinedPostType[post.user_id] = users;
+            } else {
+                const users = {ids: [], usernames: []};
+                if (userId) {
+                    users.ids.push(userId);
+                } else if (username) {
+                    users.usernames.push(username);
+                }
+                userActivityProps[postType] = {
+                    [post.user_id]: users,
+                };
+            }
+        } else {
+            const propsUserId = post.user_id;
+
+            if (combinedPostType) {
+                if (!Array.isArray(combinedPostType)) {
+                    throw new Error('Invalid Post activity data');
+                }
+                if (!combinedPostType.includes(propsUserId)) {
+                    userActivityProps[postType] = [...combinedPostType, propsUserId];
+                }
+            } else {
+                userActivityProps = {...userActivityProps, [postType]: [propsUserId]};
+            }
+        }
+
+        return userActivityProps;
+    }, {});
+
+    return extractUserActivityData(userActivities);
+}
