+++ src/utils/post_list.test.js
@@ -9,9 +9,12 @@ import {getPreferenceKey} from 'utils/preference_utils';
 
 import {
     COMBINED_USER_ACTIVITY,
+    combineUserActivitySystemPost,
+    comparePostTypes,
     DATE_LINE,
     makeCombineUserActivityPosts,
     makePreparePostIdsForPostList,
+    postTypePriority,
     START_OF_NEW_MESSAGES,
 } from './post_list';
 
@@ -780,5 +783,576 @@ describe('makeCombineUserActivityPosts', () => {
             expect(combineUserActivityPosts.recomputations()).toBe(4);
             expect(result).not.toBe(initialResult);
         });
+    });
+});
+
+describe('combineUserActivitySystemPost', () => {
+    const PostTypes = Posts.POST_TYPES;
+
+    it('should return null', () => {
+        assert.equal(Boolean(combineUserActivitySystemPost()), false);
+        assert.equal(Boolean(combineUserActivitySystemPost([])), false);
+    });
+
+    const postAddToChannel1 = {type: PostTypes.ADD_TO_CHANNEL, user_id: 'user_id_1', props: {addedUserId: 'added_user_id_1', addedUsername: 'added_username_1'}};
+    const postAddToChannel2 = {type: PostTypes.ADD_TO_CHANNEL, user_id: 'user_id_1', props: {addedUserId: 'added_user_id_2', addedUsername: 'added_username_2'}};
+    const postAddToChannel3 = {type: PostTypes.ADD_TO_CHANNEL, user_id: 'user_id_1', props: {addedUserId: 'added_user_id_3', addedUsername: 'added_username_3'}};
+    const postAddToChannel4 = {type: PostTypes.ADD_TO_CHANNEL, user_id: 'user_id_2', props: {addedUserId: 'added_user_id_4', addedUsername: 'added_username_4'}};
+    const postAddToChannel5 = {type: PostTypes.ADD_TO_CHANNEL, user_id: 'user_id_1', props: {addedUsername: 'added_username_1'}};
+    it('should match return for ADD_TO_CHANNEL', () => {
+        const out1 = {
+            allUserIds: ['added_user_id_1', 'user_id_1'],
+            allUsernames: [],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_id_1']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postAddToChannel1]), out1);
+
+        const out2 = {
+            allUserIds: ['added_user_id_1', 'added_user_id_2', 'user_id_1'],
+            allUsernames: [],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_id_1', 'added_user_id_2']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postAddToChannel1, postAddToChannel2]), out2);
+
+        const out3 = {
+            allUserIds: ['added_user_id_1', 'added_user_id_2', 'added_user_id_3', 'user_id_1'],
+            allUsernames: [],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_id_1', 'added_user_id_2', 'added_user_id_3']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postAddToChannel1, postAddToChannel2, postAddToChannel3]), out3);
+
+        const out4 = {
+            allUserIds: ['added_user_id_1', 'added_user_id_2', 'added_user_id_3', 'user_id_1', 'added_user_id_4', 'user_id_2'],
+            allUsernames: [],
+            messageData: [
+                {actorId: 'user_id_1', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_id_1', 'added_user_id_2', 'added_user_id_3']},
+                {actorId: 'user_id_2', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_id_4']},
+            ],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postAddToChannel1, postAddToChannel2, postAddToChannel3, postAddToChannel4]), out4);
+
+        const out5 = {
+            allUserIds: ['user_id_1'],
+            allUsernames: ['added_username_1'],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_username_1']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postAddToChannel5]), out5);
+
+        const out6 = {
+            allUserIds: ['added_user_id_1', 'added_user_id_2', 'added_user_id_3', 'user_id_1', 'added_user_id_4', 'user_id_2'],
+            allUsernames: ['added_username_1'],
+            messageData: [
+                {actorId: 'user_id_1', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_username_1', 'added_user_id_1', 'added_user_id_2', 'added_user_id_3']},
+                {actorId: 'user_id_2', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_id_4']},
+            ],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postAddToChannel1, postAddToChannel2, postAddToChannel3, postAddToChannel4, postAddToChannel5]), out6);
+    });
+
+    it('should match return for ADD_TO_CHANNEL, backward compatibility with addedUsername', () => {
+        const out1 = {
+            allUserIds: ['user_id_1'],
+            allUsernames: ['added_user_name_1'],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_name_1']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([{...postAddToChannel1, props: {addedUsername: 'added_user_name_1'}}]), out1);
+
+        const out2 = {
+            allUserIds: ['added_user_id_2', 'user_id_1'],
+            allUsernames: ['added_user_name_1'],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_name_1', 'added_user_id_2']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([{...postAddToChannel1, props: {addedUsername: 'added_user_name_1'}}, postAddToChannel2]), out2);
+
+        const out3 = {
+            allUserIds: ['added_user_id_1', 'added_user_id_2', 'added_user_id_3', 'user_id_1', 'user_id_2'],
+            allUsernames: ['added_user_name_4'],
+            messageData: [
+                {actorId: 'user_id_1', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_id_1', 'added_user_id_2', 'added_user_id_3']},
+                {actorId: 'user_id_2', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_name_4']},
+            ],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postAddToChannel1, postAddToChannel2, postAddToChannel3, {...postAddToChannel4, props: {addedUsername: 'added_user_name_4'}}]), out3);
+    });
+
+    const postAddToTeam1 = {type: PostTypes.ADD_TO_TEAM, user_id: 'user_id_1', props: {addedUserId: 'added_user_id_1'}};
+    const postAddToTeam2 = {type: PostTypes.ADD_TO_TEAM, user_id: 'user_id_1', props: {addedUserId: 'added_user_id_2'}};
+    const postAddToTeam3 = {type: PostTypes.ADD_TO_TEAM, user_id: 'user_id_1', props: {addedUserId: 'added_user_id_3'}};
+    const postAddToTeam4 = {type: PostTypes.ADD_TO_TEAM, user_id: 'user_id_2', props: {addedUserId: 'added_user_id_4'}};
+    it('should match return for ADD_TO_TEAM', () => {
+        const out1 = {
+            allUserIds: ['added_user_id_1', 'user_id_1'],
+            allUsernames: [],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_id_1']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postAddToTeam1]), out1);
+
+        const out2 = {
+            allUserIds: ['added_user_id_1', 'added_user_id_2', 'user_id_1'],
+            allUsernames: [],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_id_1', 'added_user_id_2']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postAddToTeam1, postAddToTeam2]), out2);
+
+        const out3 = {
+            allUserIds: ['added_user_id_1', 'added_user_id_2', 'added_user_id_3', 'user_id_1'],
+            allUsernames: [],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_id_1', 'added_user_id_2', 'added_user_id_3']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postAddToTeam1, postAddToTeam2, postAddToTeam3]), out3);
+
+        const out4 = {
+            allUserIds: ['added_user_id_1', 'added_user_id_2', 'added_user_id_3', 'user_id_1', 'added_user_id_4', 'user_id_2'],
+            allUsernames: [],
+            messageData: [
+                {actorId: 'user_id_1', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_id_1', 'added_user_id_2', 'added_user_id_3']},
+                {actorId: 'user_id_2', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_id_4']},
+            ],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postAddToTeam1, postAddToTeam2, postAddToTeam3, postAddToTeam4]), out4);
+    });
+
+    it('should match return for ADD_TO_TEAM, backward compatibility with addedUsername', () => {
+        const out1 = {
+            allUserIds: ['user_id_1'],
+            allUsernames: ['added_user_name_1'],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_name_1']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([{...postAddToTeam1, props: {addedUsername: 'added_user_name_1'}}]), out1);
+
+        const out2 = {
+            allUserIds: ['added_user_id_2', 'user_id_1'],
+            allUsernames: ['added_user_name_1'],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_name_1', 'added_user_id_2']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([{...postAddToTeam1, props: {addedUsername: 'added_user_name_1'}}, postAddToTeam2]), out2);
+
+        const out3 = {
+            allUserIds: ['added_user_id_1', 'added_user_id_2', 'added_user_id_3', 'user_id_1', 'user_id_2'],
+            allUsernames: ['added_user_name_4'],
+            messageData: [
+                {actorId: 'user_id_1', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_id_1', 'added_user_id_2', 'added_user_id_3']},
+                {actorId: 'user_id_2', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_name_4']},
+            ],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postAddToTeam1, postAddToTeam2, postAddToTeam3, {...postAddToTeam4, props: {addedUsername: 'added_user_name_4'}}]), out3);
+    });
+
+    const postJoinChannel1 = {type: PostTypes.JOIN_CHANNEL, user_id: 'user_id_1'};
+    const postJoinChannel2 = {type: PostTypes.JOIN_CHANNEL, user_id: 'user_id_2'};
+    const postJoinChannel3 = {type: PostTypes.JOIN_CHANNEL, user_id: 'user_id_3'};
+    const postJoinChannel4 = {type: PostTypes.JOIN_CHANNEL, user_id: 'user_id_4'};
+    it('should match return for JOIN_CHANNEL', () => {
+        const out1 = {
+            allUserIds: ['user_id_1'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.JOIN_CHANNEL, userIds: ['user_id_1']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postJoinChannel1]), out1);
+
+        const out2 = {
+            allUserIds: ['user_id_1', 'user_id_2'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.JOIN_CHANNEL, userIds: ['user_id_1', 'user_id_2']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postJoinChannel1, postJoinChannel2]), out2);
+
+        const out3 = {
+            allUserIds: ['user_id_1', 'user_id_2', 'user_id_3'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.JOIN_CHANNEL, userIds: ['user_id_1', 'user_id_2', 'user_id_3']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postJoinChannel1, postJoinChannel2, postJoinChannel3]), out3);
+
+        const out4 = {
+            allUserIds: ['user_id_1', 'user_id_2', 'user_id_3', 'user_id_4'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.JOIN_CHANNEL, userIds: ['user_id_1', 'user_id_2', 'user_id_3', 'user_id_4']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postJoinChannel1, postJoinChannel2, postJoinChannel3, postJoinChannel4]), out4);
+    });
+
+    const postJoinTeam1 = {type: PostTypes.JOIN_TEAM, user_id: 'user_id_1'};
+    const postJoinTeam2 = {type: PostTypes.JOIN_TEAM, user_id: 'user_id_2'};
+    const postJoinTeam3 = {type: PostTypes.JOIN_TEAM, user_id: 'user_id_3'};
+    const postJoinTeam4 = {type: PostTypes.JOIN_TEAM, user_id: 'user_id_4'};
+    it('should match return for JOIN_TEAM', () => {
+        const out1 = {
+            allUserIds: ['user_id_1'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.JOIN_TEAM, userIds: ['user_id_1']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postJoinTeam1]), out1);
+
+        const out2 = {
+            allUserIds: ['user_id_1', 'user_id_2'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.JOIN_TEAM, userIds: ['user_id_1', 'user_id_2']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postJoinTeam1, postJoinTeam2]), out2);
+
+        const out3 = {
+            allUserIds: ['user_id_1', 'user_id_2', 'user_id_3'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.JOIN_TEAM, userIds: ['user_id_1', 'user_id_2', 'user_id_3']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postJoinTeam1, postJoinTeam2, postJoinTeam3]), out3);
+
+        const out4 = {
+            allUserIds: ['user_id_1', 'user_id_2', 'user_id_3', 'user_id_4'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.JOIN_TEAM, userIds: ['user_id_1', 'user_id_2', 'user_id_3', 'user_id_4']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postJoinTeam1, postJoinTeam2, postJoinTeam3, postJoinTeam4]), out4);
+    });
+
+    const postLeaveChannel1 = {type: PostTypes.LEAVE_CHANNEL, user_id: 'user_id_1'};
+    const postLeaveChannel2 = {type: PostTypes.LEAVE_CHANNEL, user_id: 'user_id_2'};
+    const postLeaveChannel3 = {type: PostTypes.LEAVE_CHANNEL, user_id: 'user_id_3'};
+    const postLeaveChannel4 = {type: PostTypes.LEAVE_CHANNEL, user_id: 'user_id_4'};
+    it('should match return for LEAVE_CHANNEL', () => {
+        const out1 = {
+            allUserIds: ['user_id_1'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.LEAVE_CHANNEL, userIds: ['user_id_1']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postLeaveChannel1]), out1);
+
+        const out2 = {
+            allUserIds: ['user_id_1', 'user_id_2'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.LEAVE_CHANNEL, userIds: ['user_id_1', 'user_id_2']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postLeaveChannel1, postLeaveChannel2]), out2);
+
+        const out3 = {
+            allUserIds: ['user_id_1', 'user_id_2', 'user_id_3'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.LEAVE_CHANNEL, userIds: ['user_id_1', 'user_id_2', 'user_id_3']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postLeaveChannel1, postLeaveChannel2, postLeaveChannel3]), out3);
+
+        const out4 = {
+            allUserIds: ['user_id_1', 'user_id_2', 'user_id_3', 'user_id_4'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.LEAVE_CHANNEL, userIds: ['user_id_1', 'user_id_2', 'user_id_3', 'user_id_4']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postLeaveChannel1, postLeaveChannel2, postLeaveChannel3, postLeaveChannel4]), out4);
+    });
+
+    const postLeaveTeam1 = {type: PostTypes.LEAVE_TEAM, user_id: 'user_id_1'};
+    const postLeaveTeam2 = {type: PostTypes.LEAVE_TEAM, user_id: 'user_id_2'};
+    const postLeaveTeam3 = {type: PostTypes.LEAVE_TEAM, user_id: 'user_id_3'};
+    const postLeaveTeam4 = {type: PostTypes.LEAVE_TEAM, user_id: 'user_id_4'};
+    it('should match return for LEAVE_TEAM', () => {
+        const out1 = {
+            allUserIds: ['user_id_1'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.LEAVE_TEAM, userIds: ['user_id_1']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postLeaveTeam1]), out1);
+
+        const out2 = {
+            allUserIds: ['user_id_1', 'user_id_2'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.LEAVE_TEAM, userIds: ['user_id_1', 'user_id_2']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postLeaveTeam1, postLeaveTeam2]), out2);
+
+        const out3 = {
+            allUserIds: ['user_id_1', 'user_id_2', 'user_id_3'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.LEAVE_TEAM, userIds: ['user_id_1', 'user_id_2', 'user_id_3']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postLeaveTeam1, postLeaveTeam2, postLeaveTeam3]), out3);
+
+        const out4 = {
+            allUserIds: ['user_id_1', 'user_id_2', 'user_id_3', 'user_id_4'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.LEAVE_TEAM, userIds: ['user_id_1', 'user_id_2', 'user_id_3', 'user_id_4']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postLeaveTeam1, postLeaveTeam2, postLeaveTeam3, postLeaveTeam4]), out4);
+    });
+
+    const postRemoveFromChannel1 = {type: PostTypes.REMOVE_FROM_CHANNEL, user_id: 'user_id_1', props: {removedUserId: 'removed_user_id_1'}};
+    const postRemoveFromChannel2 = {type: PostTypes.REMOVE_FROM_CHANNEL, user_id: 'user_id_1', props: {removedUserId: 'removed_user_id_2'}};
+    const postRemoveFromChannel3 = {type: PostTypes.REMOVE_FROM_CHANNEL, user_id: 'user_id_1', props: {removedUserId: 'removed_user_id_3'}};
+    const postRemoveFromChannel4 = {type: PostTypes.REMOVE_FROM_CHANNEL, user_id: 'user_id_1', props: {removedUserId: 'removed_user_id_4'}};
+    it('should match return for REMOVE_FROM_CHANNEL', () => {
+        const out1 = {
+            allUserIds: ['removed_user_id_1', 'user_id_1'],
+            allUsernames: [],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.REMOVE_FROM_CHANNEL, userIds: ['removed_user_id_1']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postRemoveFromChannel1]), out1);
+
+        const out2 = {
+            allUserIds: ['removed_user_id_1', 'removed_user_id_2', 'user_id_1'],
+            allUsernames: [],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.REMOVE_FROM_CHANNEL, userIds: ['removed_user_id_1', 'removed_user_id_2']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postRemoveFromChannel1, postRemoveFromChannel2]), out2);
+
+        const out3 = {
+            allUserIds: ['removed_user_id_1', 'removed_user_id_2', 'removed_user_id_3', 'user_id_1'],
+            allUsernames: [],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.REMOVE_FROM_CHANNEL, userIds: ['removed_user_id_1', 'removed_user_id_2', 'removed_user_id_3']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postRemoveFromChannel1, postRemoveFromChannel2, postRemoveFromChannel3]), out3);
+
+        const out4 = {
+            allUserIds: ['removed_user_id_1', 'removed_user_id_2', 'removed_user_id_3', 'removed_user_id_4', 'user_id_1'],
+            allUsernames: [],
+            messageData: [{actorId: 'user_id_1', postType: PostTypes.REMOVE_FROM_CHANNEL, userIds: ['removed_user_id_1', 'removed_user_id_2', 'removed_user_id_3', 'removed_user_id_4']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postRemoveFromChannel1, postRemoveFromChannel2, postRemoveFromChannel3, postRemoveFromChannel4]), out4);
+    });
+
+    const postRemoveFromTeam1 = {type: PostTypes.REMOVE_FROM_TEAM, user_id: 'user_id_1'};
+    const postRemoveFromTeam2 = {type: PostTypes.REMOVE_FROM_TEAM, user_id: 'user_id_2'};
+    const postRemoveFromTeam3 = {type: PostTypes.REMOVE_FROM_TEAM, user_id: 'user_id_3'};
+    const postRemoveFromTeam4 = {type: PostTypes.REMOVE_FROM_TEAM, user_id: 'user_id_4'};
+    it('should match return for REMOVE_FROM_TEAM', () => {
+        const out1 = {
+            allUserIds: ['user_id_1'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.REMOVE_FROM_TEAM, userIds: ['user_id_1']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postRemoveFromTeam1]), out1);
+
+        const out2 = {
+            allUserIds: ['user_id_1', 'user_id_2'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.REMOVE_FROM_TEAM, userIds: ['user_id_1', 'user_id_2']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postRemoveFromTeam1, postRemoveFromTeam2]), out2);
+
+        const out3 = {
+            allUserIds: ['user_id_1', 'user_id_2', 'user_id_3'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.REMOVE_FROM_TEAM, userIds: ['user_id_1', 'user_id_2', 'user_id_3']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postRemoveFromTeam1, postRemoveFromTeam2, postRemoveFromTeam3]), out3);
+
+        const out4 = {
+            allUserIds: ['user_id_1', 'user_id_2', 'user_id_3', 'user_id_4'],
+            allUsernames: [],
+            messageData: [{postType: PostTypes.REMOVE_FROM_TEAM, userIds: ['user_id_1', 'user_id_2', 'user_id_3', 'user_id_4']}],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postRemoveFromTeam1, postRemoveFromTeam2, postRemoveFromTeam3, postRemoveFromTeam4]), out4);
+    });
+
+    it('should match return on combination', () => {
+        const out1 = {
+            allUserIds: ['added_user_id_1', 'added_user_id_2', 'user_id_1'],
+            allUsernames: [],
+            messageData: [
+                {actorId: 'user_id_1', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_id_1', 'added_user_id_2']},
+                {actorId: 'user_id_1', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_id_1', 'added_user_id_2']},
+            ],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postAddToChannel1, postAddToChannel2, postAddToTeam1, postAddToTeam2]), out1);
+
+        const out2 = {
+            allUserIds: ['user_id_1', 'user_id_2'],
+            allUsernames: [],
+            messageData: [
+                {postType: PostTypes.JOIN_TEAM, userIds: ['user_id_1', 'user_id_2']},
+                {postType: PostTypes.JOIN_CHANNEL, userIds: ['user_id_1', 'user_id_2']},
+            ],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postJoinChannel1, postJoinChannel2, postJoinTeam1, postJoinTeam2]), out2);
+
+        const out3 = {
+            allUserIds: ['user_id_1', 'user_id_2'],
+            allUsernames: [],
+            messageData: [
+                {postType: PostTypes.LEAVE_TEAM, userIds: ['user_id_1', 'user_id_2']},
+                {postType: PostTypes.LEAVE_CHANNEL, userIds: ['user_id_1', 'user_id_2']},
+            ],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postLeaveChannel1, postLeaveChannel2, postLeaveTeam1, postLeaveTeam2]), out3);
+
+        const out4 = {
+            allUserIds: ['removed_user_id_1', 'removed_user_id_2', 'user_id_1', 'user_id_2'],
+            allUsernames: [],
+            messageData: [
+                {postType: PostTypes.REMOVE_FROM_TEAM, userIds: ['user_id_1', 'user_id_2']},
+                {actorId: 'user_id_1', postType: PostTypes.REMOVE_FROM_CHANNEL, userIds: ['removed_user_id_1', 'removed_user_id_2']},
+            ],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([postRemoveFromChannel1, postRemoveFromChannel2, postRemoveFromTeam1, postRemoveFromTeam2]), out4);
+
+        const out5 = {
+            allUserIds: ['added_user_id_1', 'added_user_id_2', 'user_id_1', 'user_id_2', 'removed_user_id_1', 'removed_user_id_2'],
+            allUsernames: [],
+            messageData: [
+                {postType: PostTypes.JOIN_CHANNEL, userIds: ['user_id_1', 'user_id_2']},
+                {actorId: 'user_id_1', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_id_1', 'added_user_id_2']},
+                {postType: PostTypes.LEAVE_CHANNEL, userIds: ['user_id_1', 'user_id_2']},
+                {actorId: 'user_id_1', postType: PostTypes.REMOVE_FROM_CHANNEL, userIds: ['removed_user_id_1', 'removed_user_id_2']},
+            ],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([
+            postAddToChannel1,
+            postJoinChannel1,
+            postLeaveChannel1,
+            postRemoveFromChannel1,
+            postAddToChannel2,
+            postJoinChannel2,
+            postLeaveChannel2,
+            postRemoveFromChannel2,
+        ]), out5);
+
+        const out6 = {
+            allUserIds: ['added_user_id_3', 'user_id_1', 'added_user_id_4', 'user_id_2', 'user_id_3', 'user_id_4'],
+            allUsernames: [],
+            messageData: [
+                {postType: PostTypes.JOIN_TEAM, userIds: ['user_id_3', 'user_id_4']},
+                {actorId: 'user_id_1', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_id_3']},
+                {actorId: 'user_id_2', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_id_4']},
+                {postType: PostTypes.LEAVE_TEAM, userIds: ['user_id_3', 'user_id_4']},
+                {postType: PostTypes.REMOVE_FROM_TEAM, userIds: ['user_id_3', 'user_id_4']},
+            ],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([
+            postAddToTeam3,
+            postJoinTeam3,
+            postLeaveTeam3,
+            postRemoveFromTeam3,
+            postAddToTeam4,
+            postJoinTeam4,
+            postLeaveTeam4,
+            postRemoveFromTeam4,
+        ]), out6);
+
+        const out7 = {
+            allUserIds: ['added_user_id_3', 'added_user_id_1', 'added_user_id_2', 'user_id_1', 'added_user_id_4', 'user_id_2', 'user_id_3', 'user_id_4', 'removed_user_id_1', 'removed_user_id_2', 'removed_user_id_3', 'removed_user_id_4'],
+            allUsernames: [],
+            messageData: [
+                {postType: PostTypes.JOIN_TEAM, userIds: ['user_id_3', 'user_id_4', 'user_id_1', 'user_id_2']},
+                {actorId: 'user_id_1', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_id_3', 'added_user_id_1', 'added_user_id_2']},
+                {actorId: 'user_id_2', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_id_4']},
+                {postType: PostTypes.LEAVE_TEAM, userIds: ['user_id_3', 'user_id_4', 'user_id_1', 'user_id_2']},
+                {postType: PostTypes.REMOVE_FROM_TEAM, userIds: ['user_id_3', 'user_id_4', 'user_id_1', 'user_id_2']},
+                {postType: PostTypes.JOIN_CHANNEL, userIds: ['user_id_1', 'user_id_2', 'user_id_3', 'user_id_4']},
+                {actorId: 'user_id_1', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_id_1', 'added_user_id_2', 'added_user_id_3']},
+                {actorId: 'user_id_2', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_id_4']},
+                {postType: PostTypes.LEAVE_CHANNEL, userIds: ['user_id_1', 'user_id_2', 'user_id_3', 'user_id_4']},
+                {actorId: 'user_id_1', postType: PostTypes.REMOVE_FROM_CHANNEL, userIds: ['removed_user_id_1', 'removed_user_id_2', 'removed_user_id_3', 'removed_user_id_4']},
+            ],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([
+            postAddToTeam3,
+            postJoinTeam3,
+            postLeaveTeam3,
+            postRemoveFromTeam3,
+            postAddToTeam4,
+            postJoinTeam4,
+            postLeaveTeam4,
+            postRemoveFromTeam4,
+
+            postAddToChannel1,
+            postJoinChannel1,
+            postLeaveChannel1,
+            postRemoveFromChannel1,
+            postAddToChannel2,
+            postJoinChannel2,
+            postLeaveChannel2,
+            postRemoveFromChannel2,
+
+            postAddToChannel3,
+            postJoinChannel3,
+            postLeaveChannel3,
+            postRemoveFromChannel3,
+            postAddToChannel4,
+            postJoinChannel4,
+            postLeaveChannel4,
+            postRemoveFromChannel4,
+
+            postAddToTeam1,
+            postJoinTeam1,
+            postLeaveTeam1,
+            postRemoveFromTeam1,
+            postAddToTeam2,
+            postJoinTeam2,
+            postLeaveTeam2,
+            postRemoveFromTeam2,
+        ]), out7);
+
+        const out8 = {
+            allUserIds: ['added_user_id_3', 'user_id_1', 'user_id_3', 'added_user_id_1', 'removed_user_id_1'],
+            allUsernames: [],
+            messageData: [
+                {postType: PostTypes.JOIN_TEAM, userIds: ['user_id_3']},
+                {actorId: 'user_id_1', postType: PostTypes.ADD_TO_TEAM, userIds: ['added_user_id_3']},
+                {postType: PostTypes.LEAVE_TEAM, userIds: ['user_id_3']},
+                {postType: PostTypes.REMOVE_FROM_TEAM, userIds: ['user_id_3']},
+                {postType: PostTypes.JOIN_CHANNEL, userIds: ['user_id_1']},
+                {actorId: 'user_id_1', postType: PostTypes.ADD_TO_CHANNEL, userIds: ['added_user_id_1']},
+                {postType: PostTypes.LEAVE_CHANNEL, userIds: ['user_id_1']},
+                {actorId: 'user_id_1', postType: PostTypes.REMOVE_FROM_CHANNEL, userIds: ['removed_user_id_1']},
+            ],
+        };
+        assert.deepEqual(combineUserActivitySystemPost([
+            postAddToTeam3,
+            postAddToTeam3,
+            postJoinTeam3,
+            postJoinTeam3,
+            postLeaveTeam3,
+            postLeaveTeam3,
+            postRemoveFromTeam3,
+            postRemoveFromTeam3,
+
+            postAddToChannel1,
+            postAddToChannel1,
+            postJoinChannel1,
+            postJoinChannel1,
+            postLeaveChannel1,
+            postLeaveChannel1,
+            postRemoveFromChannel1,
+            postRemoveFromChannel1,
+        ]), out8);
+    });
+});
+
+describe('comparePostTypes', () => {
+    const {
+        JOIN_TEAM,
+        ADD_TO_TEAM,
+        LEAVE_TEAM,
+        REMOVE_FROM_TEAM,
+        JOIN_CHANNEL,
+        ADD_TO_CHANNEL,
+        LEAVE_CHANNEL,
+        REMOVE_FROM_CHANNEL,
+    } = Posts.POST_TYPES;
+
+    const testCases = [
+        [],
+        [{postType: JOIN_TEAM}],
+        [{postType: JOIN_TEAM}, {postType: ADD_TO_TEAM}],
+        [{postType: ADD_TO_TEAM}, {postType: JOIN_TEAM}],
+        [{postType: ADD_TO_TEAM}, {postType: ADD_TO_TEAM}, {postType: JOIN_TEAM}],
+        [{postType: JOIN_TEAM}, {postType: ADD_TO_TEAM}, {postType: LEAVE_TEAM}, {postType: REMOVE_FROM_TEAM}],
+        [{postType: REMOVE_FROM_TEAM}, {postType: LEAVE_TEAM}, {postType: ADD_TO_TEAM}, {postType: JOIN_TEAM}],
+        [{postType: JOIN_CHANNEL}, {postType: ADD_TO_CHANNEL}, {postType: LEAVE_CHANNEL}, {postType: REMOVE_FROM_CHANNEL}],
+        [{postType: REMOVE_FROM_CHANNEL}, {postType: LEAVE_CHANNEL}, {postType: ADD_TO_CHANNEL}, {postType: JOIN_CHANNEL}],
+        [{postType: LEAVE_CHANNEL}, {postType: REMOVE_FROM_CHANNEL}, {postType: LEAVE_TEAM}, {postType: REMOVE_FROM_TEAM}],
+        [{postType: LEAVE_TEAM}, {postType: REMOVE_FROM_TEAM}, {postType: LEAVE_CHANNEL}, {postType: REMOVE_FROM_CHANNEL}],
+        [{postType: JOIN_CHANNEL}, {postType: LEAVE_CHANNEL}, {postType: JOIN_CHANNEL}, {postType: REMOVE_FROM_CHANNEL}, {postType: ADD_TO_CHANNEL}],
+    ];
+
+    it('should sort post type correctly', () => {
+        for (const testCase of testCases) {
+            let previousType;
+            testCase.sort(comparePostTypes).forEach((sortedTestCase, index) => {
+                if (index > 0) {
+                    assert.ok(postTypePriority[previousType] <= postTypePriority[sortedTestCase.postType], `${previousType} should come first before ${sortedTestCase.postType}`);
+                }
+
+                previousType = sortedTestCase.postType;
+            });
+        }
     });
 });
